Fix ownership/perms on named-xs directory and rm key symlinks before we upgrade
This is a first stab at attempting a half-decent upgrade from the earlier
version. If we don't unlink the files we are about to install, the underlying
symlinks will prevent installation of the package files, which get diverted
to '.rpmnew'.

Not proud of this - but making this work is going to be complicated...