yum.conf: Attempt to replace the old yum.conf
On xs-0.4, yum.conf was put in place with a dirty
hack which means that new versions won't overwrite
it as it seems to be 'edited locally'.

So we replace it in our %post if it matches the sha1
of the version we know we want to replace. If it does
not match the sha1 we have, then it's truly a local
edit, and we respect it.