pgsql-xs: Wait for postmaster to complete startup -- solves race w Moodle
This is a function stolen from the old pg_ctl startup script taht shipped
with postgres 7.4. It defers returning success until it can connect to
the postmaster process and get it to list its DBs.

With this, Moodle DB initialisation now works, even on slow IO machines
such as the XO.

This is a highly local patch -- it works with assumptions we can make
on the XS, but are not general to a wide-use distro (see the warnings
comment block).