Re-add pgsql init script modifications
Ensure our daemon starts (chkconfig modification)
Wait for postmaster to complete startup -- solves race w Moodle

The change to initdb log location change was dropped. It's not clear
why this was included earlier, and the approach of installing logs
inside the data directory is not possible with pgsql-8.4 (it refuses
to create the db if the directory is not empty).