Rework network interface binding and firewalling
To simplify configuration, and to be resilient to working on networks
configured with layouts other than our standard one, our services now
listen on all interfaces.

For the XS standard network config case, we now use iptables to block
access to school-level services from the WAN. The new iptables rules
are based on what is generated by system-config-firewall-tui for a
similar setup.