Add xs-setup upgrade path
Some components that install setup.d scripts may need xs-setup to be
re-run upon package upgrade. It would be unsafe to unconditionally
re-run "xs-setup" from a RPM %post script, as the XS may not have
been activated.

Add a "xs-setup --upgrade-only" path to solve this situation.
In this case, xs-setup will do nothing unless xs-setup was previously
and successfully run in the past.